requirements.txt
pandas>=1.4.0
numpy>=1.22.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
pyzmq>=23.0.0
joblib>=1.1.0
